backend/requirements.txt
fastapi==0.121.0
uvicorn[standard]==0.38.0
sqlmodel==0.0.22
alembic==1.13.3
pydantic-settings==2.11.0
httpx==0.28.1
python-multipart==0.0.20
openpyxl==3.1.5
pillow==11.3.0
pandas==2.2.3

# 测试依赖
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0

